Create Status path
